refactor: conversations

+++ src/list/handlers/callback/delete.ts
@@ -1,10 +1,14 @@
 import { CallbackQueryContext } from "grammy";
 import { CustomGeneralContext } from "../../../app/shared/interfaces";
-import { mainActionsKeyboard } from "../../../app/shared/keyboards/main-actions.keyboard";
 import { deleteOne } from "../../list.service";
+import { handleShowListsInlineKeyboard } from "../../keyboards/lists.inline-keyboard";
 
 export const deleteCallback = async (ctx: CallbackQueryContext<CustomGeneralContext>) => {
   const listId = ctx.match[1];
+
+  const telegramId = ctx.chat?.id
+
+  if(!telegramId) return;
 
   const deleteTodo = await deleteOne({
     where: {
@@ -12,7 +16,9 @@ export const deleteCallback = async (ctx: CallbackQueryContext<CustomGeneralCont
     },
   });
 
+// todo: MAKE CONVO
+
   return ctx.reply(`${deleteTodo.name} - deleted!`, {
-    reply_markup: mainActionsKeyboard,
+    reply_markup:await handleShowListsInlineKeyboard(telegramId),
   });
 };
